requirements.txt
anthropic>=0.25.0
openai>=1.3.0
dashscope>=1.14.0
mysql-connector-python>=8.2.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
